ref(issues): Remove unused AssigneTo component and reorganize assignee selector utils
