+++ static/app/components/feedback/feedbackItem/feedbackAssignedTo.tsx
@@ -2,8 +2,6 @@ import {useEffect} from 'react';
 
 import {fetchOrgMembers} from 'sentry/actionCreators/members';
 import {useFeedbackCache} from 'sentry/components/feedback/useFeedbackCache';
-import type {EventOwners} from 'sentry/components/group/assignedTo';
-import {getOwnerList} from 'sentry/components/group/assignedTo';
 import {
   AssigneeSelector,
   useHandleAssigneeChange,
@@ -14,6 +12,8 @@ import type {FeedbackEvent} from 'sentry/utils/feedback/types';
 import {useApiQuery} from 'sentry/utils/queryClient';
 import {useApi} from 'sentry/utils/useApi';
 import {useOrganization} from 'sentry/utils/useOrganization';
+import type {EventOwners} from 'sentry/views/issueDetails/streamline/header/getOwnerList';
+import {getOwnerList} from 'sentry/views/issueDetails/streamline/header/getOwnerList';
 
 interface Props {
   feedbackEvent: FeedbackEvent;
