+++ static/app/utils/useIssueEventOwners.tsx
@@ -1,8 +1,8 @@
-import type {EventOwners} from 'sentry/components/group/assignedTo';
 import {getApiUrl} from 'sentry/utils/api/getApiUrl';
 import type {ApiQueryKey, UseApiQueryOptions} from 'sentry/utils/queryClient';
 import {useApiQuery} from 'sentry/utils/queryClient';
 import {useOrganization} from 'sentry/utils/useOrganization';
+import type {EventOwners} from 'sentry/views/issueDetails/streamline/header/getOwnerList';
 
 interface UseIssueEventOwnersProps {
   eventId: string;
