+++ static/app/views/issueDetails/streamline/header/assigneeSelector.spec.tsx
@@ -8,10 +8,10 @@ import {UserFixture} from 'sentry-fixture/user';
 
 import {render, screen, userEvent} from 'sentry-test/reactTestingLibrary';
 
-import type {EventOwners} from 'sentry/components/group/assignedTo';
 import {MemberListStore} from 'sentry/stores/memberListStore';
 import type {Committer} from 'sentry/types/integrations';
 import {GroupHeaderAssigneeSelector} from 'sentry/views/issueDetails/streamline/header/assigneeSelector';
+import type {EventOwners} from 'sentry/views/issueDetails/streamline/header/getOwnerList';
 
 describe('GroupHeaderAssigneeSelector', () => {
   const organization = OrganizationFixture();
