+++ static/app/views/issueDetails/streamline/header/assigneeSelector.tsx
@@ -5,7 +5,6 @@ import {MenuComponents} from '@sentry/scraps/compactSelect';
 
 import {fetchOrgMembers} from 'sentry/actionCreators/members';
 import {openIssueOwnershipRuleModal} from 'sentry/actionCreators/modal';
-import {getOwnerList} from 'sentry/components/group/assignedTo';
 import {
   AssigneeSelector,
   useHandleAssigneeChange,
@@ -19,6 +18,7 @@ import {useApi} from 'sentry/utils/useApi';
 import {useCommitters} from 'sentry/utils/useCommitters';
 import {useIssueEventOwners} from 'sentry/utils/useIssueEventOwners';
 import {useOrganization} from 'sentry/utils/useOrganization';
+import {getOwnerList} from 'sentry/views/issueDetails/streamline/header/getOwnerList';
 
 interface GroupHeaderAssigneeSelectorProps {
   event: Event | null;
